base-hw: implement quota scheduler with background priorities
This is a second scheduler which uses the same
method to scheduler claims, but a choose a priority based scheduling for the
fills. The process with the highest priority gets scheduled first and
only schedules a different process, if a process with higher priority is
available or the current process changes it's state to unready.