TRACE: make native thread ids available via trace_session
As we use native thread ids for core trace events, we need the
possibility to relate these ids to the corresponding thread names.
Note that there are two kinds of threads traced by core events:
those created via a CPU session (which also appear as trace subjects)
and those created by core.

We make native thread ids available via two mechanisms.
Both rely on the id() function that we added to Platform_thread.
The first mechanism is an extension to the Subject_info and hence only
covers threads created via a CPU session.
The second mechanism is the newly added core_info() function of the
trace_session, which provides an XML for all known threads.

relates to #2317