trace: eagerly init trace buffer
The buffer should be initialised before the TRACE client tries to read
any entries from it. Lazily initialising the buffer when it is attached
by the traced thread, however, requires the TRACE client to wait for the
state change. Instead, we can already initialise the buffer when it is
allocated and avoid any coordination protocol.

genodelabs/genode#4434