nic_router: avoid marginal timeout updates
The link dissolve timeout is updated for every packet, which leads to
trigger_once() RPCs that only marginally change the scheduled timeout but
significantly slow down the packet throughput.

Instead of updating the link timeout immediately, we only reschedule the
timeout either once the old timeout fired or if the old timeout would be
sufficiently near in the future (50% of the intended duration).