event_filter: make multitouch tracking more robust
When the multitouch driver occasionally looses track of a finger, it
may occur that touch id 0 is released before the other ids. This messes
with the gesture detection. This commit adds a Multitouch struct that reliably
tracks the presence of fingers for all gestures and thereby moves this state
out of the individual gestures.