hw: re-order x86 Vcpu destruction correctly
This commit primarily addresses the wrong destruction order of
Core::Vm_session_component::Vcpu for x86_64, which was destroying
the Vcpu hardware buffer before calling the kernel to invalidate its
own Kernel::Vcpu object that depends on this hardware buffer.

To facilitate the re-ordering, but especially to enhance maintainability
for the future, and to stay more consistent with other architectures,
most part of this commit re-works how the Vcpu hardware buffer (VMCB,
or VMCS) is represented in core/kernel. When x86 virtualization got
introduced a Genode::Vcpu_data class was introduced therefore.
Since then it polluted the common Genode public API across all
architectures and components, although this implementation detail
is relevant for the hw-kernel on x86 only. Therefore, this commit
removes Vcpu_data completely.
Instead, the architecture-specific Board::Vcpu_state, which is mostly
just the Vcpu state description shared between hypervisor and VMM,
gets extended in the Board::Vcpu_state case for x86_64 to contain
the VMM state, as well as the hardware buffers.

Fix #5628